+++ example/rtl/main.js
@@ -1,0 +1,23 @@
+'use strict';
+
+// Create an instance
+var wavesurfer;
+
+// Init & load audio file
+document.addEventListener('DOMContentLoaded', function() {
+    // Init
+    wavesurfer = WaveSurfer.create({
+        container: document.querySelector('#waveform'),
+        waveColor: '#A8DBA8',
+        progressColor: '#3B8686',
+        backend: 'MediaElement',
+        rtl: true
+    });
+
+    // Load audio from URL
+    wavesurfer.load('../media/demo.wav');
+
+    document
+        .querySelector('[data-action="play"]')
+        .addEventListener('click', wavesurfer.playPause.bind(wavesurfer));
+});