+++ src/wavesurfer.js
@@ -29,6 +29,8 @@ import PeakCache from './peakcache';
  * @property {string} backend='WebAudio' `'WebAudio'|'MediaElement'` In most cases
  * you don't have to set this manually. MediaElement is a fallback for
  * unsupported browsers.
+ * @property {string} backgroundColor=null Change background color of the
+ * waveform container.
  * @property {number} barHeight=1 The height of the wave
  * @property {number} barGap=null The optional spacing between bars of the wave,
  * if not provided will be calculated in legacy format.
@@ -39,6 +41,8 @@ import PeakCache from './peakcache';
  * @property {string} cursorColor='#333' The fill color of the cursor indicating
  * the playhead position.
  * @property {number} cursorWidth=1 Measured in pixels.
+ * @property {number} duration=null Optional audio length so pre-rendered peaks
+ * can be display immediately for example.
  * @property {boolean} fillParent=true Whether to fill the entire container or
  * draw only according to `minPxPerSec`.
  * @property {boolean} forceDecode=false Force decoding of audio using web audio
@@ -82,6 +86,8 @@ import PeakCache from './peakcache';
  * @property {boolean|number} responsive=false If set to `true` resize the
  * waveform, when the window is resized. This is debounced with a `100ms`
  * timeout by default. If this parameter is a number it represents that timeout.
+ * @property {boolean} rtl=false If set to `true`, renders waveform from
+ * right-to-left.
  * @property {boolean} scrollParent=false Whether to scroll the container with a
  * lengthy waveform. Otherwise the waveform is shrunk to the container width
  * (see fillParent).
@@ -182,12 +188,14 @@ export default class WaveSurfer extends util.Observer {
         audioRate: 1,
         autoCenter: true,
         backend: 'WebAudio',
+        backgroundColor: null,
         barHeight: 1,
         barGap: null,
         container: null,
         cursorColor: '#333',
         cursorWidth: 1,
         dragSelection: true,
+        duration: null,
         fillParent: true,
         forceDecode: false,
         height: 128,
@@ -208,6 +216,7 @@ export default class WaveSurfer extends util.Observer {
         removeMediaElementOnDestroy: true,
         renderer: MultiCanvas,
         responsive: false,
+        rtl: false,
         scrollParent: false,
         skipLength: 2,
         splitChannels: false,
@@ -311,6 +320,14 @@ export default class WaveSurfer extends util.Observer {
             throw new Error('maxCanvasWidth must be greater than 1');
         } else if (this.params.maxCanvasWidth % 2 == 1) {
             throw new Error('maxCanvasWidth must be an even number');
+        }
+
+        if (this.params.rtl === true) {
+            util.style(this.container, { transform: 'rotateY(180deg)' });
+        }
+
+        if (this.params.backgroundColor) {
+            this.setBackgroundColor(this.params.backgroundColor);
         }
 
         /**
@@ -1004,6 +1021,26 @@ export default class WaveSurfer extends util.Observer {
     }
 
     /**
+     * Get the background color of the waveform container.
+     *
+     * @return {string} A CSS color string.
+     */
+    getBackgroundColor() {
+        return this.params.backgroundColor;
+    }
+
+    /**
+     * Set the background color of the waveform container.
+     *
+     * @param {string} color A CSS color string.
+     * @example wavesurfer.setBackgroundColor('#FF00FF');
+     */
+    setBackgroundColor(color) {
+        this.params.backgroundColor = color;
+        util.style(this.container, { background: this.params.backgroundColor });
+    }
+
+    /**
      * Get the fill color of the cursor indicating the playhead
      * position.
      *
@@ -1060,7 +1097,8 @@ export default class WaveSurfer extends util.Observer {
         );
         const parentWidth = this.drawer.getWidth();
         let width = nominalWidth;
-        let start = this.drawer.getScrollX();
+        // always start at 0 after zooming for scrolling : issue redraw left part
+        let start = 0;
         let end = Math.max(start + parentWidth, width);
         // Fill container
         if (