+++ CHANGES.md
@@ -1,6 +1,37 @@
 wavesurfer.js changelog
 =======================
 
+2.2.1 (unreleased)
+------------------
+
+- Add `backgroundColor` option (#1118)
+- Spectrogram plugin: fix click handler (#1585)
+- Cursor plugin: fix `displayTime` (#1589)
+
+2.2.0 (07.03.2019)
+------------------
+
+- Add `rtl` option (#1296)
+- Fix peaks rendering issue on zooming and scrolling multicanvas (#1570)
+- Add `duration` option to specify an explicit audio length (#1441)
+- Spectrogram plugin: fix event listener removal (#1571)
+- Regions plugin: display regions before file load using `duration`
+  option (#1441)
+- Build: switch to terser-webpack-plugin for minifying
+
+2.1.3 (21.01.2019)
+------------------
+
+- Fix removeOnAudioProcess for Safari (#1215, #1367, #1398)
+
+2.1.2 (06.01.2019)
+------------------
+
+- Fix computing peaks when buffer is not set (#1530)
+- Cursor plugin: fix displayed time (#1543)
+- Cursor plugin: document new params (#1516)
+- Add syntax highlighting in examples (#1522)
+
 2.1.1 (18.11.2018)
 ------------------